+ Modified to BigInteger

+++ src/main/kotlin/App.kt
@@ -1,4 +1,5 @@
 import java.io.File
+import java.math.BigInteger
 
 fun readFileLines(fileName: String): List<String> {
     val file = File(fileName)
@@ -9,7 +10,7 @@ fun readFileLines(fileName: String): List<String> {
 fun main(args: Array<String>) {
     val input = readFileLines("in.txt")
     val parser = Parser()
-    val output = ArrayList<Long>()
+    val output = ArrayList<BigInteger>()
     for (i in 0..input.lastIndex)
         output.add(parser.parse(input[i]).value)
     File("out.txt").printWriter().use { out ->
