+++ src/main/kotlin/tokens/PrimaryInteger.kt
@@ -1,5 +1,7 @@
 package tokens
 
-class PrimaryInteger(value: Long): Primary() {
+import java.math.BigInteger
+
+class PrimaryInteger(value: BigInteger): Primary() {
     override var value = value
 }
