+++ src/main/kotlin/tokens/Relation.kt
@@ -1,23 +1,25 @@
 package tokens
 
+import java.math.BigInteger
+
 open class Relation: Binary() {
 
     class Less(left: Expression, right: Expression) : Relation() {
         override var left = left
         override var right = right
-        override var value: Long = if (left.value < right.value) 1L else 0L
+        override var value: BigInteger = if (left.value < right.value) BigInteger.valueOf(1) else BigInteger.valueOf(0)
     }
 
     class More(left: Expression, right: Expression): Relation() {
         override var left = left
         override var right = right
-        override var value: Long = if (left.value > right.value) 1L else 0L
+        override var value: BigInteger = if (left.value > right.value) BigInteger.valueOf(1) else BigInteger.valueOf(0)
     }
 
     class Equal(left: Expression, right: Expression): Relation() {
         override var left = left
         override var right = right
-        override var value: Long = if (left.value == right.value) 1L else 0L
+        override var value: BigInteger = if (left.value == right.value) BigInteger.valueOf(1) else BigInteger.valueOf(0)
     }
 
 }
