+++ src/main/kotlin/tokens/Term.kt
@@ -1,17 +1,19 @@
 package tokens
 
+import java.math.BigInteger
+
 open class Term: Binary() {
 
     class Plus(left: Expression,right: Expression): Term() {
         override var left = left
         override var right = right
-        override var value: Long = left.value + right.value
+        override var value: BigInteger = left.value + right.value
     }
 
     class Minus(left: Expression,right: Expression): Term() {
         override var left = left
         override var right = right
-        override var value: Long = left.value - right.value
+        override var value: BigInteger = left.value - right.value
     }
 
 }
